move mida_trainer to multi_domain_adapter.py and add one hot encoder inside the trainer
