+++ docs/source/kale.pipeline.rst
@@ -30,14 +30,6 @@ kale.pipeline.domain\_adapter module
    :undoc-members:
    :show-inheritance:
 
-kale.pipeline.mida\_trainer module
-----------------------------------
-
-.. automodule:: kale.pipeline.mida_trainer
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 kale.pipeline.mpca\_trainer module
 ----------------------------------
