+++ kale/evaluate/cross_validation.py
@@ -36,7 +36,7 @@ def _fit_and_score(
     y,
     transformer,
     domain_adapter,
-    factors,
+    group_labels,
     scorer,
     train,
     test,
@@ -54,7 +54,7 @@ def _fit_and_score(
     error_score=np.nan,
 ):
     """Train the estimator (optionally with a domain adapter) and evaluate its performance.
-    The implementation is a modification of the `scikit-learn`'s _fit_and_score function to accomodate
+    The implementation is a modification of the `scikit-learn`'s _fit_and_score function to accommodate
     domain adaptation and pre-domain adaptation transformation. Do not use this function with classification
     labels that has -1 as a label, as it will be treated as an unlabeled sample. This is a limitation of the
     current implementation.
@@ -65,8 +65,8 @@ def _fit_and_score(
         y (array-like): Target variable for supervised learning [num_samples] or [num_samples, num_targets].
         transformer (sklearn.base.BaseEstimator, optional): An unsupervised transformer implementing fit and transform methods applied before domain adaptation.
         domain_adapter (sklearn.base.BaseEstimator, optional): A domain adapter implementing fit and transform methods.
-        factors (array-like): The factors for adaptation with shape (num_samples, num_factors). Please preprocess the factors before domain adaptation
-                            (e.g. one-hot encode domain, gender, or standardize age).
+        group_labels (array-like): Categorical variables representing domain or grouping factors with shape
+            (num_samples, num_factors).
         scorer (callable): A scoring function to evaluate the estimator's performance.
         train (array-like): Indices of training samples.
         test (array-like): Indices of testing samples.
@@ -114,9 +114,9 @@ def _fit_and_score(
 
     if verbose > 2:
         if split_progress is not None:
-            progress_msg = f" {split_progress[0]+1}/{split_progress[1]}"
+            progress_msg = f" {split_progress[0] + 1}/{split_progress[1]}"
         if candidate_progress and verbose > 9:
-            progress_msg += f"; {candidate_progress[0]+1}/{candidate_progress[1]}"
+            progress_msg += f"; {candidate_progress[0] + 1}/{candidate_progress[1]}"
 
     if verbose > 1:
         if parameters is None:
@@ -169,9 +169,9 @@ def _fit_and_score(
 
     if domain_adapter is not None:
         factors_sampled, factors_train, factors_test = [None] * 3
-        if factors is not None:
-            factors_train = _safe_indexing(factors, train)
-            factors_test = _safe_indexing(factors, test)
+        if group_labels is not None:
+            factors_train = _safe_indexing(group_labels, train)
+            factors_test = _safe_indexing(group_labels, test)
             factors_sampled = xp.concatenate([factors_train, factors_test], axis=0)
 
         if parameters is not None:
