+++ tests/embed/test_factorization.py
@@ -107,10 +107,10 @@ def test_mida_shape_consistency(sample_data, num_components):
 
     mida = MIDA(num_components=num_components)
     mida.set_params(**mida.get_params())
-    mida.fit(x, factors=factors)
+    mida.fit(x, group_labels=factors)
 
     # Transform the whole data
-    z = mida.transform(x, factors=factors)
+    z = mida.transform(x, group_labels=factors)
 
     # If num_components is not None, check the shape of the transformed data
     if num_components is not None:
@@ -118,8 +118,8 @@ def test_mida_shape_consistency(sample_data, num_components):
 
     # Transform the source and target domain data separately
     (source_mask,) = np.where(domains != 0)
-    z_src = mida.transform(x[source_mask], factors=factors[source_mask])
-    z_tgt = mida.transform(x[~source_mask], factors=factors[~source_mask])
+    z_src = mida.transform(x[source_mask], group_labels=factors[source_mask])
+    z_tgt = mida.transform(x[~source_mask], group_labels=factors[~source_mask])
     # Check if transformations are consistent with separate domains
     testing.assert_allclose(z_src, z[source_mask])
     testing.assert_allclose(z_tgt, z[~source_mask])
@@ -134,10 +134,10 @@ def test_mida_inverse_transform(sample_data):
     x, y, domains, factors = sample_data
 
     mida = MIDA(fit_inverse_transform=True)
-    mida.fit(x, factors=factors)
+    mida.fit(x, group_labels=factors)
 
     # Transform the whole data
-    z = mida.transform(x, factors=factors)
+    z = mida.transform(x, group_labels=factors)
     # Inverse transform the data
     x_rec = mida.inverse_transform(z)
 
@@ -154,7 +154,7 @@ def test_mida_support_kernel(sample_data, kernel):
     x, y, domains, factors = sample_data
 
     mida = MIDA(num_components=N_COMP_CONSTANT, kernel=kernel)
-    mida.fit(x, factors=factors)
+    mida.fit(x, group_labels=factors)
 
     is_linear = kernel == "linear"
     try:
@@ -164,7 +164,7 @@ def test_mida_support_kernel(sample_data, kernel):
         assert not is_linear, "MIDA must not have `orig_coef_` after fitting when kernel!='linear'"
 
     # Transform the whole data
-    z = mida.transform(x, factors=factors)
+    z = mida.transform(x, group_labels=factors)
 
     # expect to allow multiple kernels supported
     assert mida._n_features_out == N_COMP_CONSTANT, f"Expected {N_COMP_CONSTANT} components, got {mida._n_features_out}"
@@ -176,7 +176,7 @@ def test_mida_augment(sample_data, augment):
     x, y, domains, factors = sample_data
 
     mida = MIDA(num_components=N_COMP_CONSTANT, kernel="linear", augment=augment, fit_inverse_transform=True)
-    mida.fit(x, factors=factors)
+    mida.fit(x, group_labels=factors)
 
     # expect validator for domain factors if augment=True
     # has_validator = hasattr(mida, "_factor_validator")
@@ -187,7 +187,7 @@ def test_mida_augment(sample_data, augment):
     assert mida._factor_validator is None, "MIDA must not have `_factor_validator` after fitting when augment=False"
 
     if augment == "post":
-        z = mida.transform(x, factors=factors)
+        z = mida.transform(x, group_labels=factors)
         assert (
             z.shape[1] == N_COMP_CONSTANT + factors.shape[-1]
         ), f"Expected {x.shape[1]} features, got {N_COMP_CONSTANT + factors.shape[-1]}"
@@ -198,7 +198,7 @@ def test_mida_ignore_y(sample_data, ignore_y):
     x, y, domains, factors = sample_data
 
     mida = MIDA(num_components=N_COMP_CONSTANT, kernel="linear", ignore_y=ignore_y)
-    mida.fit(x, y, factors=factors)
+    mida.fit(x, y, group_labels=factors)
 
     # expect classes_ to be set if ignore_y=False
     # has_classes = hasattr(mida, "classes_")
@@ -219,10 +219,10 @@ def test_mida_eigen_solver(sample_data, eigen_solver):
         eigen_solver=eigen_solver,
         max_iter=200 if eigen_solver == "arpack" else None,
     )
-    mida.fit(x, factors=factors)
+    mida.fit(x, group_labels=factors)
 
     # Transform the whole data
-    z = mida.transform(x, factors=factors)
+    z = mida.transform(x, group_labels=factors)
 
     # expect the solver to have a consistent number of components
     assert mida._n_features_out == N_COMP_CONSTANT, f"Expected {N_COMP_CONSTANT} components, got {mida._n_features_out}"
@@ -234,10 +234,10 @@ def test_mida_scale_components(sample_data, scale_components):
     x, y, domains, factors = sample_data
 
     mida = MIDA(num_components=N_COMP_CONSTANT, kernel="linear", scale_components=scale_components)
-    mida.fit(x, factors=factors)
+    mida.fit(x, group_labels=factors)
 
     # Transform the whole data
-    z = mida.transform(x, factors=factors)
+    z = mida.transform(x, group_labels=factors)
 
     # Expect the scale_components to have a consistent number of components
     # the behavior expected is the zero eigenvalues component is masked, not indexed
